refactor(ipc): D4 — split connector-rpc-handlers.ts into namespace directory
Replaces the 1106-LOC packages/gateway/src/ipc/connector-rpc-handlers.ts
with a directory of namespace-focused sibling files:

- context.ts    — ConnectorRpcHandlerContext + ConnectorRpcHit types
- status.ts     — listStatus / status / healthHistory
- lifecycle.ts  — pause / resume / sync + shared internal helpers
- config.ts     — addMcp / setInterval / setConfig (imports lifecycle helpers)
- removal.ts    — remove / resumePendingRemovals + 5 removal helpers
- auth.ts       — auth + 18 per-connector flows + OAuth helpers
- index.ts      — pure re-export shim preserving the original public surface

Pure mechanical move; zero behavioral change. Function bodies migrate
verbatim; per-file imports trim to minimum scope. Three consumer files
(connector-rpc.ts, platform/assemble.ts, connector-rpc-handlers-setconfig
.test.ts) update their import path string from
"./connector-rpc-handlers.ts" to "./connector-rpc-handlers/index.ts".

The largest resulting file is auth.ts at 676 LOC, well under the 800 D4
threshold.

Spec: docs/superpowers/specs/2026-05-02-d4-rpc-handlers-split-design.md
Plan: docs/superpowers/plans/2026-05-02-d4-rpc-handlers-split.md
