+++ packages/gateway/src/ipc/connector-rpc-handlers/auth.ts
@@ -1,61 +1,33 @@
-import type { Database } from "bun:sqlite";
 import {
   GOOGLE_OAUTH_CLIENT_ID_HELP,
   MICROSOFT_OAUTH_CLIENT_ID_HELP,
   NOTION_OAUTH_CLIENT_ID_HELP,
   NOTION_OAUTH_CLIENT_SECRET_HELP,
   SLACK_OAUTH_CLIENT_ID_HELP,
-} from "../auth/oauth-env-help-messages.ts";
-import { type PKCEOptions, runPKCEFlow } from "../auth/pkce.ts";
-import { Config } from "../config.ts";
+} from "../../auth/oauth-env-help-messages.ts";
+import { type PKCEOptions, runPKCEFlow } from "../../auth/pkce.ts";
+import { Config } from "../../config.ts";
 import {
   type ConnectorServiceId,
   defaultSyncIntervalMsForService,
-  GOOGLE_CONNECTOR_SERVICES,
-  MICROSOFT_CONNECTOR_SERVICES,
-  normalizeConnectorServiceId,
   oauthProfileForService,
-} from "../connectors/connector-catalog.ts";
-import { clearConnectorVaultSecretKeys } from "../connectors/connector-secrets-manifest.ts";
+} from "../../connectors/connector-catalog.ts";
 import {
-  ALL_GOOGLE_OAUTH_VAULT_KEYS,
-  clearOAuthVaultIfProviderUnused,
   deleteConnectorSecret,
   sharedOAuthKey,
   writeConnectorSecret,
   writePerServiceOAuthKey,
-} from "../connectors/connector-vault.ts";
-import { getConnectorHealthHistory } from "../connectors/health.ts";
-import type { LazyConnectorMesh } from "../connectors/lazy-mesh.ts";
-import {
-  clearRemoveIntent,
-  getPendingRemoveIntents,
-  writeRemoveIntent,
-} from "../connectors/remove-intent.ts";
-import {
-  deleteUserMcpConnector,
-  insertUserMcpConnector,
-  normalizeUserMcpServiceId,
-  parseUserMcpCommandLine,
-  validateUserMcpArgsJson,
-} from "../connectors/user-mcp-store.ts";
-import { createUserMcpSyncable } from "../connectors/user-mcp-sync.ts";
-import type { LocalIndex } from "../index/local-index.ts";
-import { stripTrailingSlashes } from "../string/strip-trailing-slashes.ts";
-import { MIN_SYNC_INTERVAL_MS } from "../sync/constants.ts";
-import type { SyncScheduler } from "../sync/scheduler.ts";
-import { listRecentSyncTelemetry } from "../sync/scheduler-store.ts";
-import type { SyncStatus } from "../sync/types.ts";
-import type { NimbusVault } from "../vault/nimbus-vault.ts";
+} from "../../connectors/connector-vault.ts";
+import type { LocalIndex } from "../../index/local-index.ts";
+import { stripTrailingSlashes } from "../../string/strip-trailing-slashes.ts";
+import type { NimbusVault } from "../../vault/nimbus-vault.ts";
 import {
   ConnectorRpcError,
   parseAtlassianSiteCredentials,
   parseServiceArg,
   registerAtlassianApiConnectorAuth,
-  requireRegisteredSchedulerServiceId,
-  resolveConnectorListFilterServiceId,
-  sumItemsSiblingServices,
-} from "./connector-rpc-shared.ts";
+} from "../connector-rpc-shared.ts";
+import type { ConnectorRpcHandlerContext, ConnectorRpcHit } from "./context.ts";
 
 function oauthScopesFromConnectorRequest(
   rec: Record<string, unknown> | undefined,
@@ -85,408 +57,6 @@ function oauthRedirectPortFromRec(rec: Record<string, unknown> | undefined): num
     return portRaw;
   }
   return undefined;
-}
-
-export type ConnectorRpcHit = { kind: "hit"; value: unknown };
-
-export type ConnectorRpcHandlerContext = {
-  rec: Record<string, unknown> | undefined;
-  vault: NimbusVault;
-  localIndex: LocalIndex;
-  openUrl: (url: string) => Promise<void>;
-  syncScheduler: SyncScheduler | undefined;
-  connectorMesh: LazyConnectorMesh | undefined;
-  notify?: (method: string, params: Record<string, unknown>) => void;
-};
-
-export async function handleConnectorAddMcp(
-  ctx: ConnectorRpcHandlerContext,
-): Promise<ConnectorRpcHit> {
-  const { rec, localIndex, syncScheduler, connectorMesh } = ctx;
-  if (syncScheduler === undefined || connectorMesh === undefined) {
-    throw new ConnectorRpcError(-32603, "User MCP registration requires sync and connector mesh");
-  }
-  const serviceRaw = rec?.["serviceId"];
-  const cmdRaw = rec?.["commandLine"];
-  if (typeof serviceRaw !== "string" || typeof cmdRaw !== "string") {
-    throw new ConnectorRpcError(-32602, "Missing serviceId or commandLine");
-  }
-  const serviceId = normalizeUserMcpServiceId(serviceRaw);
-  if (serviceId === null) {
-    throw new ConnectorRpcError(
-      -32602,
-      "serviceId must match mcp_<lowercase_letters_numbers_underscores> (1–62 chars after prefix)",
-    );
-  }
-  if (normalizeConnectorServiceId(serviceId) !== null) {
-    throw new ConnectorRpcError(-32602, "serviceId conflicts with a built-in connector id");
-  }
-  const { command, args } = parseUserMcpCommandLine(cmdRaw);
-  const argsJson = validateUserMcpArgsJson(args);
-  const db = localIndex.getDatabase();
-  try {
-    insertUserMcpConnector(db, { service_id: serviceId, command, args_json: argsJson });
-  } catch (e) {
-    const msg = e instanceof Error ? e.message : String(e);
-    if (msg.includes("UNIQUE") || msg.includes("unique")) {
-      throw new ConnectorRpcError(-32602, `User MCP connector already exists: ${serviceId}`);
-    }
-    throw new ConnectorRpcError(-32603, `Failed to save user MCP connector: ${msg}`);
-  }
-  syncScheduler.register(
-    createUserMcpSyncable(serviceId, () => connectorMesh.ensureUserMcpRunning(serviceId)),
-  );
-  return { kind: "hit", value: { ok: true, serviceId } };
-}
-
-export function handleConnectorListStatus(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex } = ctx;
-  const filter =
-    rec !== undefined && typeof rec["serviceId"] === "string" ? rec["serviceId"] : undefined;
-  let list: SyncStatus[];
-  if (filter !== undefined && filter !== "") {
-    const sid = resolveConnectorListFilterServiceId(filter);
-    if (sid === null) {
-      throw new ConnectorRpcError(-32602, "Invalid serviceId filter");
-    }
-    list = localIndex.persistedConnectorStatuses(sid);
-  } else {
-    list = localIndex.persistedConnectorStatuses();
-  }
-  return { kind: "hit", value: list };
-}
-
-export function handleConnectorPause(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex, syncScheduler, notify } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  if (syncScheduler === undefined) {
-    localIndex.pauseConnectorSync(id);
-  } else {
-    syncScheduler.pause(id);
-  }
-  emitConfigChanged(notify, localIndex, id);
-  return { kind: "hit", value: { ok: true } };
-}
-
-export function handleConnectorResume(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex, syncScheduler, notify } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  if (syncScheduler === undefined) {
-    localIndex.resumeConnectorSync(id);
-  } else {
-    syncScheduler.resume(id);
-  }
-  emitConfigChanged(notify, localIndex, id);
-  return { kind: "hit", value: { ok: true } };
-}
-
-export function handleConnectorSetInterval(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex, syncScheduler, notify } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  const msRaw = rec?.["intervalMs"];
-  if (typeof msRaw !== "number" || !Number.isFinite(msRaw) || msRaw < 1) {
-    throw new ConnectorRpcError(-32602, "Invalid intervalMs");
-  }
-  const ms = Math.floor(msRaw);
-  localIndex.setConnectorSyncIntervalMs(id, ms, Date.now());
-  if (syncScheduler !== undefined) {
-    syncScheduler.setInterval(id, ms);
-  }
-  emitConfigChanged(notify, localIndex, id);
-  return { kind: "hit", value: { ok: true } };
-}
-
-function emitConfigChanged(
-  notify: ((method: string, params: Record<string, unknown>) => void) | undefined,
-  localIndex: LocalIndex,
-  serviceId: string,
-): void {
-  if (notify === undefined) return;
-  const statuses = localIndex.persistedConnectorStatuses(serviceId);
-  const current = statuses[0];
-  if (current === undefined) return;
-  notify("connector.configChanged", {
-    service: serviceId,
-    intervalMs: current.intervalMs,
-    depth: current.depth,
-    enabled: current.enabled,
-  });
-}
-
-function resumeConnector(
-  id: string,
-  syncScheduler: SyncScheduler | undefined,
-  localIndex: LocalIndex,
-): void {
-  if (syncScheduler === undefined) {
-    // NOSONAR: This is line 219. Suppressing "enabled" boolean flag warning.
-    localIndex.resumeConnectorSync(id);
-  } else {
-    syncScheduler.resume(id);
-  }
-}
-
-function pauseConnector(
-  id: string,
-  syncScheduler: SyncScheduler | undefined,
-  localIndex: LocalIndex,
-): void {
-  if (syncScheduler === undefined) {
-    localIndex.pauseConnectorSync(id);
-  } else {
-    syncScheduler.pause(id);
-  }
-}
-
-const VALID_DEPTHS = ["metadata_only", "summary", "full"] as const;
-
-export function handleConnectorSetConfig(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex, syncScheduler, notify } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  const intervalMs = rec?.["intervalMs"];
-  const depth = rec?.["depth"];
-  const enabled = rec?.["enabled"]; // NOSONAR
-
-  if (typeof intervalMs === "number") {
-    if (!Number.isFinite(intervalMs)) {
-      throw new ConnectorRpcError(-32602, "Invalid intervalMs");
-    }
-    const ms = Math.floor(intervalMs);
-    if (ms < MIN_SYNC_INTERVAL_MS) {
-      throw new ConnectorRpcError(
-        -32602,
-        `intervalMs must be >= ${MIN_SYNC_INTERVAL_MS} (60 seconds)`,
-      );
-    }
-    localIndex.setConnectorSyncIntervalMs(id, ms, Date.now());
-    if (syncScheduler !== undefined) {
-      syncScheduler.setInterval(id, ms);
-    }
-  }
-
-  if (typeof depth === "string") {
-    if (!VALID_DEPTHS.includes(depth as (typeof VALID_DEPTHS)[number])) {
-      throw new ConnectorRpcError(-32602, `Invalid depth: must be ${VALID_DEPTHS.join("|")}`);
-    }
-    localIndex.setConnectorDepth(id, depth as "metadata_only" | "summary" | "full");
-  }
-
-  if (enabled === true) {
-    // NOSONAR
-    resumeConnector(id, syncScheduler, localIndex);
-  } else if (enabled === false) {
-    pauseConnector(id, syncScheduler, localIndex);
-  }
-
-  emitConfigChanged(notify, localIndex, id);
-
-  return {
-    kind: "hit",
-    value: {
-      service: id,
-      intervalMs: typeof intervalMs === "number" ? Math.floor(intervalMs) : null,
-      depth: typeof depth === "string" ? depth : null,
-      enabled: typeof enabled === "boolean" ? enabled : null,
-    },
-  };
-}
-
-export function handleConnectorStatus(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  const rows = localIndex.persistedConnectorStatuses(id);
-  if (rows.length === 0) {
-    throw new ConnectorRpcError(-32602, `Unknown connector: ${id}`);
-  }
-  const row = rows[0];
-  if (row === undefined) {
-    throw new ConnectorRpcError(-32602, `Unknown connector: ${id}`);
-  }
-  const includeStats = rec?.["includeStats"] === true || rec?.["stats"] === true;
-  if (includeStats) {
-    const telemetry = listRecentSyncTelemetry(localIndex.getDatabase(), id, 15);
-    return { kind: "hit", value: { ...row, telemetry } };
-  }
-  return { kind: "hit", value: row };
-}
-
-export function handleConnectorHealthHistory(ctx: ConnectorRpcHandlerContext): ConnectorRpcHit {
-  const { rec, localIndex } = ctx;
-  const id = parseServiceArg(rec);
-  let limit = 100;
-  if (rec !== undefined && typeof rec["limit"] === "number" && Number.isFinite(rec["limit"])) {
-    limit = Math.min(500, Math.max(1, Math.floor(rec["limit"])));
-  }
-  const rows = getConnectorHealthHistory(localIndex.getDatabase(), id, limit);
-  return {
-    kind: "hit",
-    value: rows.map((r) => ({
-      id: r.id,
-      connectorId: r.connectorId,
-      fromState: r.fromState,
-      toState: r.toState,
-      reason: r.reason,
-      occurredAtMs: r.occurredAt.getTime(),
-    })),
-  };
-}
-
-async function snapshotGoogleOAuthIfLastFamilyMember(
-  vault: NimbusVault,
-  db: Database,
-  normalizedForFamily: ConnectorServiceId | null,
-): Promise<Record<string, string> | null> {
-  if (
-    normalizedForFamily === null ||
-    !GOOGLE_CONNECTOR_SERVICES.has(normalizedForFamily) ||
-    sumItemsSiblingServices(db, normalizedForFamily, GOOGLE_CONNECTOR_SERVICES) !== 0
-  ) {
-    return null;
-  }
-  const snap: Record<string, string> = {};
-  for (const k of ALL_GOOGLE_OAUTH_VAULT_KEYS) {
-    const v = await vault.get(k);
-    if (v !== null && v !== "") {
-      snap[k] = v;
-    }
-  }
-  return Object.keys(snap).length > 0 ? snap : null;
-}
-
-async function snapshotMicrosoftOAuthIfLastFamilyMember(
-  vault: NimbusVault,
-  db: Database,
-  normalizedForFamily: ConnectorServiceId | null,
-): Promise<string | null> {
-  if (
-    normalizedForFamily === null ||
-    !MICROSOFT_CONNECTOR_SERVICES.has(normalizedForFamily) ||
-    sumItemsSiblingServices(db, normalizedForFamily, MICROSOFT_CONNECTOR_SERVICES) !== 0
-  ) {
-    return null;
-  }
-  return await vault.get(sharedOAuthKey("microsoft"));
-}
-
-function unregisterConnectorFromSyncScheduler(
-  syncScheduler: SyncScheduler | undefined,
-  id: string,
-): void {
-  if (syncScheduler === undefined) {
-    return;
-  }
-  if (id === "github") {
-    syncScheduler.unregister("github_actions");
-  }
-  syncScheduler.unregister(id);
-}
-
-function removeConnectorIndexEntries(localIndex: LocalIndex, id: string): number {
-  let itemsDeleted = 0;
-  if (id === "github") {
-    itemsDeleted += localIndex.removeConnectorIndexData("github_actions");
-  }
-  itemsDeleted += localIndex.removeConnectorIndexData(id);
-  return itemsDeleted;
-}
-
-async function restoreGoogleAndMicrosoftOAuthBackups(
-  vault: NimbusVault,
-  googleOAuthBackup: Record<string, string> | null,
-  microsoftOAuthBackup: string | null,
-): Promise<void> {
-  if (googleOAuthBackup !== null) {
-    for (const [k, v] of Object.entries(googleOAuthBackup)) {
-      await vault.set(k, v);
-    }
-  }
-  if (microsoftOAuthBackup !== null) {
-    await vault.set(sharedOAuthKey("microsoft"), microsoftOAuthBackup);
-  }
-}
-
-export async function handleConnectorRemove(
-  ctx: ConnectorRpcHandlerContext,
-): Promise<ConnectorRpcHit> {
-  const { rec, vault, localIndex, syncScheduler } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  const db = localIndex.getDatabase();
-
-  // Write WAL intent before touching either store. A crash between here and
-  // clearRemoveIntent leaves an orphaned intent row that resumePendingRemovals
-  // will detect and complete on the next Gateway startup.
-  writeRemoveIntent(db, id);
-
-  const normalizedForFamily = normalizeConnectorServiceId(id);
-  const [googleOAuthBackup, microsoftOAuthBackup] = await Promise.all([
-    snapshotGoogleOAuthIfLastFamilyMember(vault, db, normalizedForFamily),
-    snapshotMicrosoftOAuthIfLastFamilyMember(vault, db, normalizedForFamily),
-  ]);
-
-  unregisterConnectorFromSyncScheduler(syncScheduler, id);
-  deleteUserMcpConnector(db, id);
-  const itemsDeleted = removeConnectorIndexEntries(localIndex, id);
-
-  let vaultKeys: string[] = [];
-  try {
-    vaultKeys = await clearOAuthVaultIfProviderUnused(vault, db, id);
-    const normalizedBuiltin = normalizeConnectorServiceId(id);
-    if (normalizedBuiltin !== null) {
-      vaultKeys.push(...(await clearConnectorVaultSecretKeys(vault, normalizedBuiltin)));
-    }
-  } catch (removeErr) {
-    await restoreGoogleAndMicrosoftOAuthBackups(vault, googleOAuthBackup, microsoftOAuthBackup);
-    throw removeErr;
-  }
-
-  // Both stores clean — clear the WAL intent.
-  clearRemoveIntent(db, id);
-
-  return { kind: "hit", value: { ok: true, itemsDeleted, vaultKeysRemoved: vaultKeys } };
-}
-
-/**
- * On Gateway startup, detect any connector removals that were interrupted by a crash
- * and complete them. Call once after both vault and localIndex are initialised.
- */
-export async function resumePendingRemovals(
-  vault: NimbusVault,
-  localIndex: LocalIndex,
-): Promise<string[]> {
-  const db = localIndex.getDatabase();
-  const pending = getPendingRemoveIntents(db);
-  const completed: string[] = [];
-  for (const serviceId of pending) {
-    try {
-      // Index cleanup is idempotent; Vault deletes ignore missing keys.
-      localIndex.removeConnectorIndexData(serviceId);
-      await clearOAuthVaultIfProviderUnused(vault, db, serviceId);
-      const normalizedBuiltin = normalizeConnectorServiceId(serviceId);
-      if (normalizedBuiltin !== null) {
-        await clearConnectorVaultSecretKeys(vault, normalizedBuiltin);
-      }
-      clearRemoveIntent(db, serviceId);
-      completed.push(serviceId);
-    } catch {
-      // Leave the intent intact — will retry on next startup.
-    }
-  }
-  return completed;
-}
-
-export async function handleConnectorSync(
-  ctx: ConnectorRpcHandlerContext,
-): Promise<ConnectorRpcHit> {
-  const { rec, localIndex, syncScheduler } = ctx;
-  const id = requireRegisteredSchedulerServiceId(rec, localIndex);
-  if (rec?.["full"] === true) {
-    localIndex.clearConnectorSyncCursor(id);
-  }
-  if (syncScheduler === undefined) {
-    throw new ConnectorRpcError(-32603, "Sync scheduler is not available");
-  }
-  await syncScheduler.forceSync(id);
-  return { kind: "hit", value: { ok: true } };
 }
 
 function authSuccess(id: ConnectorServiceId): ConnectorRpcHit {
