+++ packages/gateway/src/ipc/connector-rpc-handlers/index.ts
@@ -1,0 +1,18 @@
+export { handleConnectorAuth } from "./auth.ts";
+export {
+  handleConnectorAddMcp,
+  handleConnectorSetConfig,
+  handleConnectorSetInterval,
+} from "./config.ts";
+export type { ConnectorRpcHandlerContext } from "./context.ts";
+export {
+  handleConnectorPause,
+  handleConnectorResume,
+  handleConnectorSync,
+} from "./lifecycle.ts";
+export { handleConnectorRemove, resumePendingRemovals } from "./removal.ts";
+export {
+  handleConnectorHealthHistory,
+  handleConnectorListStatus,
+  handleConnectorStatus,
+} from "./status.ts";
